requirements.txt
accelerate==1.6.0
datasets==3.5.1
deepspeed==0.16.7
easydict==1.13
einops==0.8.1
fastapi==0.116.1
func-timeout==4.3.5
jaxtyping==0.3.2
matplotlib==3.10.1
numpy==2.2.6
nvidia-ml-py==13.590.48
pandas==2.2.3
psutil==7.0.0
pydantic==2.11.7
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
submitit==1.5.3
termcolor==3.1.0
torch==2.7.1
tqdm==4.67.1
transformers==4.55.2
uvicorn==0.35.0
vllm==0.10.1
wandb==0.19.10
